extract re-used code into common module
