+++ test/utils/index.ts
@@ -1,4 +1,7 @@
 import Lumino from 'lumino-js-sdk';
+import { LuminoTestEnvironment } from '../../src/types/lumino-test-environment';
+import { LuminoNode } from '../../src/types/node';
+import { ChannelState } from '../common';
 
 export function isEmpty(value: any): boolean {
     return Array.isArray(value) ? value.length <= 0 : !value;
@@ -8,32 +11,14 @@ export function sleep(ms: number) {
     return new Promise((resolve) => setTimeout(resolve, ms));
 }
 
-export enum State {
-    OPEN = 'opened',
-    CLOSED = 'closed',
-}
-
-export class ChannelState {
-    /* Representation of channel data as obtained from a Lumino API. */
-    token: string; // hex address
-    partner: string; // hex address
-    deposit: number; // in Wei
-    balance: number; // in Wei
-    state: State;
-
-    constructor(
-        tokenAddr: string,
-        partnerAddr: string,
-        totalDeposit: number,
-        balance: number,
-        state: State
-    ) {
-        this.token = tokenAddr;
-        this.partner = partnerAddr;
-        this.deposit = totalDeposit;
-        this.balance = balance;
-        this.state = state;
-    }
+export async function verifyEnv(env: LuminoTestEnvironment): Promise<void> {
+    expect(Object.keys(env.nodes).length).toBe(2);
+    const firstNode = env.nodes.node0 as LuminoNode;
+    const secondNode = env.nodes.node1 as LuminoNode;
+    const firstAddr = (await firstNode.client.sdk.getAddress()).our_address;
+    const secondAddr = (await secondNode.client.sdk.getAddress()).our_address;
+    expect(isEmpty(firstAddr)).toBe(false);
+    expect(isEmpty(secondAddr)).toBe(false);
 }
 
 export async function verifyChannel(
